requirements.txt
flask
mysql-connector-python
google-generativeai
requests
python-dotenv
flask-cors